Change output formats a bit
Adds the ability for a slack check to say which channel the message
should be output to.

Adds configuration options for an erb message template that can be used
to parse / render event data.
* This is added as a configuration entry under the slack token.
* Adds a dependancy to erubis

Adds enumeration of client keys to add as 'fields' in the slack posting
* This is added as an array of entries under 'field' in the slack configuration.
* This arbritarily defines anything < 50 characters as "short" and slack will render them side by side.

Adds a title to the slack message to make things prettier

Helper function to translate status codes to english text.